CARE #67: Display simple bar charts

+++ src/test/java/org/motechproject/carereporting/web/chart/builder/AxisBuilderTest.java
@@ -41,7 +41,7 @@ public class AxisBuilderTest {
 
     @Test
     public void testMin() {
-        int min = -3;
+        double min = -3;
 
         axisBuilder.min(min);
